requirements.txt
django==3.2
psycopg2-binary==2.8.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.6.0
django-simple-history==2.12.0
django-extensions==3.1.0
django-cacheops==5.1
django-redis==4.12.1
django-storages==1.9.1
dateutils==0.6.12
django-filter==2.4.0
requests==2.25.1
drf-yasg==1.20.0
rest-condition==1.0.3
Pillow==8.3.2
PyJWT==2.1.0
